refactored into a single package, added setup script
