+++ .gitignore
@@ -57,6 +57,9 @@ hs_err_pid* # virtual machine crash logs, see http://www.java.com/en/download/he
 # python 
 **/*.pyc
 **/__pycache__
+*.egg-info
+dist/
+build/
 
 # shared folders
 dropbox_shared*/
