+++ setup.py
@@ -1,0 +1,10 @@
+import setuptools
+
+setuptools.setup(
+    name="dtriac-wikification",
+    version="0.0.1",
+    author="Brandeis Lab for Linguistics and Computation",
+    description="Slicing wikiump and performing document-level wikication",
+    url="https://not.yet.available.io",
+    packages=['wikification']
+)
