+++ wikification/slice/enwiki-cirrus-to-title.sh
